requirements.txt
gradio
pandas
numpy
matplotlib
fpdf
python-docx